incresce size

+++ Assets/points.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class points : MonoBehaviour
 {
-    public int pointsCount;
+    public float pointsCount;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
